ptraf: add support for inet_sock_set_state
Add BPF tracepoint to be notified when sockets are created.
It allows to add them to the list even if they don't send
actual data.

For now, the behavior is to send a SockMsgEvent with zero as len.
This can probably be improved, so we can monitor the state
and state transitions of the sockets, and keep them alive
event if they don't send anything.